+++ compiler/src/CLI_errors.ml
@@ -57,4 +57,4 @@ let check_options () =
   if !latexfile <> ""
   then warning Deprecated Location.i_dummy
          "the [-latex] option has been deprecated since March 2023; use [jazz2tex] instead";
-  List.iter chk_out_file [ outfile; latexfile; ecfile ]
+  List.iter chk_out_file [ outfile; latexfile; ecfile; json_stack_alloc_file ]